Move existing preflights to new module
