+++ cumulusci/cumulusci.yml
@@ -169,15 +169,15 @@ tasks:
         group: Data Operations
     get_installed_packages:
         description: Retrieves a list of the currently installed managed package namespaces and their versions
-        class_path: cumulusci.tasks.salesforce.GetInstalledPackages
+        class_path: cumulusci.tasks.preflight.packages.GetInstalledPackages
         group: Salesforce Preflight Checks
     get_available_licenses:
         description: Retrieves a list of the currently available license definition keys
-        class_path: cumulusci.tasks.salesforce.license_preflights.GetAvailableLicenses
+        class_path: cumulusci.tasks.preflight.licenses.GetAvailableLicenses
         group: Salesforce Preflight Checks
     get_available_permission_set_licenses:
         description: Retrieves a list of the currently available Permission Set License definition keys
-        class_path: cumulusci.tasks.salesforce.license_preflights.GetAvailablePermissionSetLicenses
+        class_path: cumulusci.tasks.preflight.licenses.GetAvailablePermissionSetLicenses
         group: Salesforce Preflight Checks
     github_parent_pr_notes:
         description: Merges the description of a child pull request to the respective parent's pull request (if one exists).
